Return complete list of MD5 sums from checksum.
Before this patch, the method almost correctly computed the list of MD5 sums but returned the result of calling paths.each, not the md5sums list.  Now paths.collect.flatten is the return value.

The find(1) call was changed slightly to use the %P formatter so the output lines look like this:

	0123456789abcdef0123456789abcdef path/to/file

Instead of this:

	0123456789abcdef0123456789abcdef ./path/to/file

The difference is cosmetic but looks like everyone else's packages.