deb: Always provide a "changes" file
lintian in Ubuntu 14.04 now errors when a file
/usr/share/NAME/changelog.Debian.gz doesn't exist.

In the spirit of respecting lintian's meaningful errors, this commit
fixes that error.

The new behavior is that a changelog file is *always* provided. If the
--deb-changelog flag is not given, then fpm will produce a default one
based on the package being built in hope that it satisfies lintian.

With this commit, the lintian rspec test passes on Ubuntu 14.04.

Fixes #784